+++ src/assets/icons/icons.ts
@@ -110,4 +110,5 @@ export default {
   location_inactive: require('./location_inactive.png'),
   plot_inactive: require('./plot_inactive.png'),
   pin: require('./pin.png'),
+  drone_auto: require('./drone-auto.png'),
 };