+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Enhance [1.4.21] 2023-09-26
+
+- action sheet confirm booking
+
 ## Improvements [v1.4.20] 2023-09-18
 
 - improve code MA
@@ -15,13 +19,15 @@
 - improve maintenance
 
 ## Improvements [1.4.16] 2023-09-04
+
 - improve code
 - text scale fix
-- notification pagination 
+- notification pagination
 - main screen split component
--  loading skeleton 
-- progressive image 
+- loading skeleton
+- progressive image
 - myJob screen pagination
+
 ## Enhance [1.4.15] 2023-08-30
 
 - improve code